Increase precision of `scale_factor` in `Windowed`

+++ native/src/window/renderer.rs
@@ -41,7 +41,7 @@ pub trait Target {
         window: &W,
         width: u32,
         height: u32,
-        scale_factor: f32,
+        scale_factor: f64,
         renderer: &Self::Renderer,
     ) -> Self;
 
@@ -52,7 +52,7 @@ pub trait Target {
         &mut self,
         width: u32,
         height: u32,
-        scale_factor: f32,
+        scale_factor: f64,
         renderer: &Self::Renderer,
     );
 }
